requirements.txt
streamlit
PyMuPDF
Pillow
google-cloud-vision
python-docx